Added support of network applications fuzzing via UDP
This patch allows AFL to do fuzzing of a target that receives data
via UDP protocol.

To be able to reduce amount of changes in WinAFL we handle recfrom
function the same way WinAFL already handles CreateFile. So, the
idea of this patch is simple: we ask WinAFL to perform required
mutation on input and then send this input over network in the
target application.

Thus requirements for the target function hasn't changed:
1. The function (or subsequent functions called from original function)
calls recvfrom.
2. The function (or subsequent functions) parses received data.
3. The function normally returns.
In addition we have one new requirement:
4. An application executes the target function in a loop to be able to
receive next incoming packages from the network.

Since our target function is executed in a loop we don't need to
interfere execution flow. However -fuzz_iterations is still used to
restart the target after specified number of iterations and need to
be specified by user.

This patch introduces three new options: -a <ip_address> -p <port> -w
<delay_in_milliseconds).

-w option is required to give the target application time to initialize
everything and open a socket for data receiving. Otherwise WinAFL might
send a packet even before the target is ready to receive that packet
thereby messing the whole process.